# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ErrorPayload": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey"
      },
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "in": "cookie",
        "name": "yauth_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "OpenAPI 3.1 specification for the yauth-go authentication library. Routes are mounted by each plugin under whatever prefix the embedder chooses (commonly `/api/auth`).",
    "license": {
      "name": "MIT"
    },
    "title": "yauth-go",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/oauth/register": {
      "post": {
        "description": "Public dynamic client registration endpoint. Disabled by default; opt in via plugin Config.DCREnabled. When DCRRequireInitialAccessToken is true (default), the request must carry an admin-issued Bearer token in the Authorization header.",
        "operationId": "oauth2DynamicClientRegister",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Oauth2RegisterRequest"
              }
            }
          },
          "description": "Client metadata per RFC 7591 §2",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Oauth2RegisterResponse"
                }
              }
            },
            "description": "Registered client + one-time client_secret (omitted for public clients) + registration_access_token."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid client metadata."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid initial access token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "DCR is disabled by configuration."
          }
        },
        "security": [],
        "summary": "Register an OAuth2 client (RFC 7591)",
        "tags": [
          "oauth2-server"
        ]
      }
    },
...
    "/oauth/{provider}": {
      "delete": {
        "operationId": "oauthUnlink",
        "parameters": [
          {
            "description": "Registered provider name (e.g. google, github).",
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Unlinked."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not authenticated."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No link for that provider."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Removing this would leave the caller with no auth method."
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Unlink a provider; refuses if it would lock out the caller",
        "tags": [
          "oauth"
        ]
      }
    },
...
  },
...
}